[#55] Update the cypress tests by using the new integration test library
