+++ integrationTests/.gitignore
@@ -1,0 +1,3 @@
+node_modules/
+cypress/screenshots/
+cypress/videos/
