--- integrationTests/cypress/fixtures/ces_admin_data.json
@@ -1,5 +1,0 @@
-{
-  "username": "ces-admin",
-  "password": "ecosystem2016",
-  "admingroup": "CesAdministrators"
-}
