+++ integrationTests/cypress/integration/cas_browser.feature
@@ -1,27 +1,27 @@
 Feature: Browser-based CAS login and logout functionality
 
   Scenario: logged out user is redirected to CAS login page
-    Given the admin is logged out of the CES
-    When the user opens the redmine dogu start page
+    Given the current user is logged out of the CES
+    When the user opens the dogu start page
     Then the user is redirected to the CAS login page
 
   Scenario: logged out user can log in to the dogu
-    Given the admin is logged out of the CES
-    When the user opens the redmine dogu start page
+    Given the current user is logged out of the CES
+    When the user opens the dogu start page
     And the user types in correct login credentials
     And the user presses the login button
     Then the user is logged in to the dogu
 
   Scenario: logged out user can not log in to the dogu with wrong credentials
-    Given the admin is logged out of the CES
-    When the user opens the redmine dogu start page
+    Given the current user is logged out of the CES
+    When the user opens the dogu start page
     And the user types in wrong login credentials
     And the user presses the login button
-    Then the login page informs user about invalid credentials
+    Then the login page informs the user about invalid credentials
 
   Scenario: logged in user can log out via logout button (front-channel)
     Given the admin is logged in to the CES
-    When the user opens the redmine dogu start page
+    When the user opens the dogu start page
     And the user clicks the logout button
     Then the user is logged out of the dogu
