+++ integrationTests/cypress/integration/redmine_cas_rest.feature
@@ -11,7 +11,8 @@ Feature: Redmine CAS plugin rest API functionality
     Then the user receives a 401 response
 
   Scenario: user can authenticate with valid api key
-    Given the user has a valid api key
+    Given the user has valid login credentials
+    And the user has a valid api key
     When the user authenticate via api key
     Then the user receives a json response with valid cas attributes
