+++ integrationTests/cypress/support/index.js
@@ -1,22 +1,6 @@
-// ***********************************************************
-// This example support/index.js is processed and
-// loaded automatically before your test files.
-//
-// This is a great place to put global configuration and
-// behavior that modifies Cypress.
-//
-// You can change the location of this file or turn off
-// automatically serving support files with the
-// 'supportFile' configuration option.
-//
-// You can read more here:
-// https://on.cypress.io/configuration
-// ***********************************************************
+// Loads all commands from the dogu integration library into this project
+const doguTestLibrary = require('@cloudogu/dogu-integration-test-library')
+doguTestLibrary.registerCommands()
 
-// Import commands.js using ES2015 syntax:
-import './commands/misc_commands'
-import './commands/redmine_commands'
-import './commands/usermgt_commands'
-
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
+// local commands
+import './commands/redmine_gui'
