+++ integrationTests/cypress/support/step_definitions/dogu_integration_test_library_steps.js
@@ -1,0 +1,3 @@
+// Loads all steps from the dogu integration library into this project
+const doguTestLibrary = require('@cloudogu/dogu-integration-test-library')
+doguTestLibrary.registerSteps()
